refactor: centralize dtype handling via DtypePolicy
Fix the test-time dtype mismatch (Half input vs float32 conv) by making
TrainerParams.precision the single source of truth for dtype resolution.

- Add dynvision/utils/dtype_policy.py: canonical PRECISION_TO_DTYPE map
  (stored-parameter semantics), resolve_dtype (warn + float32 default),
  DtypePolicy dataclass, and shared coordinate_component_dtypes helper.
- Route get_effective_dtype_from_precision and DtypeDeviceCoordinator
  through the shared resolver; replace silent float16 fallbacks with
  warn + float32.
- Delete DataParams.precision and dead ModelParams.target_dtype.
- Apply the same coordinate_component_dtypes validator to TestingParams
  (removing the 'precision' alias that swallowed unscoped precision and
  caused data to derive float16).
- Default dataloaders to float32.

Regression tests in tests/params/test_dtype_policy.py.

+++ dynvision/data/dataloader.py
@@ -74,7 +74,7 @@ class StandardDataLoader(DataLoader):
         *args,
         n_timesteps: int = 1,
         memory_format: torch.memory_format = torch.contiguous_format,
-        dtype: torch.dtype = torch.float16,
+        dtype: torch.dtype = torch.float32,
         device: Optional[str] = None,
         use_channels_last: bool = False,
         use_cuda_streams: bool = True,
